+++ common/ImpSourceVolume.h
@@ -76,7 +76,7 @@
 ///Identify which blocks are marked as allocated in the volume bitmap but have not been read from, and print those to the log.
 - (void) reportBlocksThatAreAllocatedButHaveNotBeenAccessed;
 ///Count how many blocks are marked as allocated in the volume bitmap but have not been read from. Use this method after all files have been copied when identifying orphaned blocks for recovery.
-- (NSUInteger) numberOfBlocksThatAreAllocatedButHaveNotBeenAccessed;
+- (u_int32_t) numberOfBlocksThatAreAllocatedButHaveNotBeenAccessed;
 ///Call the block with an NSRange containing each contiguous extent of blocks that are marked as allocated in the volume bitmap but have not been read from. Use this method after all files have been copied when identifying orphaned blocks for recovery.
 - (void) findExtentsThatAreAllocatedButHaveNotBeenAccessed:(void (^_Nonnull const)(NSRange))block;
 
@@ -89,7 +89,7 @@
 - (u_int32_t) firstPhysicalBlockOfFirstAllocationBlock;
 ///The offset in bytes at which the first allocation block begins. (I.e., firstPhysicalBlockOfFirstAllocationBlock converted to a byte offset.)
 - (off_t) offsetOfFirstAllocationBlock;
-- (NSUInteger) numberOfBytesPerBlock;
+- (u_int32_t) numberOfBytesPerBlock;
 ///The total number of allocation blocks in the volume, according to the volume header.
 - (NSUInteger) numberOfBlocksTotal;
 - (NSUInteger) numberOfBlocksUsed;
